train_spdc getting bug fixed
