+++ train_spdc.py
@@ -2,11 +2,123 @@ from argparse import ArgumentParser
 import yaml
 
 import torch
+import numpy as np
 from models import FNO2d
 from train_utils import Adam
 from train_utils.datasets import SPDCLoader
-from train_utils.train_2d import train_SPDC
-from train_utils.eval_2d import eval_burgers
+from train_utils.utils import save_checkpoint
+from train_utils.losses import LpLoss, darcy_loss, PINO_loss, SPDC_loss
+from tqdm import tqdm
+
+try:
+    import wandb
+except ImportError:
+    wandb = None
+
+def train_SPDC(model,
+                    train_loader, 
+                    optimizer, scheduler,
+                    config,
+                    rank=0, log=False,
+                    project='PINO-2d-default',
+                    group='default',
+                    tags=['default'],
+                    use_tqdm=True):
+    if rank == 0 and wandb and log:
+        run = wandb.init(project=project,
+                         entity=config['log']['entity'],
+                         group=group,
+                         config=config,
+                         tags=tags, reinit=True,
+                         settings=wandb.Settings(start_method="fork"))
+
+    data_weight = config['train']['xy_loss']
+    f_weight = config['train']['f_loss']
+    ic_weight = config['train']['ic_loss']
+    model.train()
+    pbar = range(config['train']['epochs'])
+    if use_tqdm:
+        pbar = tqdm(pbar, dynamic_ncols=True, smoothing=0.1)
+
+    for e in pbar:
+        model.train()
+        train_pino = 0.0
+        data_l2 = 0.0
+        train_loss = 0.0
+
+        for x, y in train_loader:
+            x, y = x.to(rank), y.to(rank)
+            out = model(x).reshape(y.shape)
+            data_loss,ic_loss,f_loss = SPDC_loss(out,x,y)
+            total_loss = ic_loss * ic_weight + f_loss * f_weight + data_loss * data_weight
+
+            optimizer.zero_grad()
+            total_loss.backward()
+            optimizer.step()
+
+            data_l2 += data_loss.item()
+            train_pino += f_loss.item()
+            train_loss += total_loss.item()
+        scheduler.step()
+        data_l2 /= len(train_loader)
+        train_pino /= len(train_loader)
+        train_loss /= len(train_loader)
+        if use_tqdm:
+            pbar.set_description(
+                (
+                    f'Epoch {e}, train loss: {train_loss:.5f} '
+                    f'train f error: {train_pino:.5f}; '
+                    f'data l2 error: {data_l2:.5f}'
+                )
+            )
+        if wandb and log:
+            wandb.log(
+                {
+                    'Train f error': train_pino,
+                    'Train L2 error': data_l2,
+                    'Train loss': train_loss,
+                }
+            )
+
+        if e % 100 == 0:
+            save_checkpoint(config['train']['save_dir'],
+                            config['train']['save_name'].replace('.pt', f'_{e}.pt'),
+                            model, optimizer)
+    save_checkpoint(config['train']['save_dir'],
+                    config['train']['save_name'],
+                    model, optimizer)
+    print('Done!')
+
+
+def eval_SPDC(model,
+                 dataloader,
+                 config,
+                 device,
+                 use_tqdm=True):
+    model.eval()
+    if use_tqdm:
+        pbar = tqdm(dataloader, dynamic_ncols=True, smoothing=0.05)
+    else:
+        pbar = dataloader
+
+    test_err = []
+    f_err = []
+
+    for x, y in pbar:
+        x, y = x.to(device), y.to(device)
+        out = model(x).reshape(y.shape)
+        data_loss,ic_loss,f_loss = SPDC_loss(out,x,y)
+        test_err.append(data_loss.item())
+        f_err.append(f_loss.item())
+
+    mean_f_err = np.mean(f_err)
+    std_f_err = np.std(f_err, ddof=1) / np.sqrt(len(f_err))
+
+    mean_err = np.mean(test_err)
+    std_err = np.std(test_err, ddof=1) / np.sqrt(len(test_err))
+
+    print(f'==Averaged relative L2 error mean: {mean_err}, std error: {std_err}==\n'
+          f'==Averaged equation error mean: {mean_f_err}, std error: {std_f_err}==')
 
 
 def run(args, config):
@@ -16,7 +128,6 @@ def run(args, config):
                             nx=data_config['nx'], 
                             ny=data_config['ny'],
                             nz=data_config['nz'],
-                            sub=data_config['sub'], 
                             F = data_config['F'],
                             datapath2 = None,
                             sub_xy=data_config['sub_xy'],
@@ -26,7 +137,8 @@ def run(args, config):
                                        batch_size=config['train']['batchsize'],
                                        start=data_config['offset'])
 
-    model = FNO2d(modes1=config['model']['modes1'],
+    model = FNO2d(in_dim=121, # should check more carfully
+                  modes1=config['model']['modes1'],
                   modes2=config['model']['modes2'],
                   fc_dim=config['model']['fc_dim'],
                   layers=config['model']['layers'],
@@ -59,7 +171,6 @@ def test(config):
                             nx=data_config['nx'], 
                             ny=data_config['ny'],
                             nz=data_config['nz'],
-                            sub=data_config['sub'], 
                             F = data_config['F'],
                             datapath2 = None,
                             sub_xy=data_config['sub_xy'],
@@ -79,7 +190,7 @@ def test(config):
         ckpt = torch.load(ckpt_path)
         model.load_state_dict(ckpt['model'])
         print('Weights loaded from %s' % ckpt_path)
-    eval_burgers(model, dataloader, dataset.v, config, device) # need to create and update to eval_SPDC
+    eval_SPDC(model, dataloader, config, device)
 
 
 if __name__ == '__main__':
