+++ train_utils/train_2d.py
@@ -2,7 +2,7 @@ import numpy as np
 import torch
 from tqdm import tqdm
 from .utils import save_checkpoint
-from .losses import LpLoss, darcy_loss, PINO_loss, SPDC_loss
+from .losses import LpLoss, darcy_loss, PINO_loss
 
 try:
     import wandb
@@ -116,79 +116,6 @@ def train_2d_operator(model,
     print('Done!')
 
 
-def train_SPDC(model,
-                    train_loader, 
-                    optimizer, scheduler,
-                    config,
-                    rank=0, log=False,
-                    project='PINO-2d-default',
-                    group='default',
-                    tags=['default'],
-                    use_tqdm=True):
-    if rank == 0 and wandb and log:
-        run = wandb.init(project=project,
-                         entity=config['log']['entity'],
-                         group=group,
-                         config=config,
-                         tags=tags, reinit=True,
-                         settings=wandb.Settings(start_method="fork"))
-
-    data_weight = config['train']['xy_loss']
-    f_weight = config['train']['f_loss']
-    ic_weight = config['train']['ic_loss']
-    model.train()
-    pbar = range(config['train']['epochs'])
-    if use_tqdm:
-        pbar = tqdm(pbar, dynamic_ncols=True, smoothing=0.1)
-
-    for e in pbar:
-        model.train()
-        train_pino = 0.0
-        data_l2 = 0.0
-        train_loss = 0.0
-
-        for x, y in train_loader:
-            x, y = x.to(rank), y.to(rank)
-            out = model(x).reshape(y.shape)
-            data_loss,ic_loss,pde_loss = SPDC_loss(out,x,y)
-            total_loss = ic_loss * ic_weight + pde_loss * f_weight + data_loss * data_weight
-
-            optimizer.zero_grad()
-            total_loss.backward()
-            optimizer.step()
-
-            data_l2 += data_loss.item()
-            train_pino += loss_f.item()
-            train_loss += total_loss.item()
-        scheduler.step()
-        data_l2 /= len(train_loader)
-        train_pino /= len(train_loader)
-        train_loss /= len(train_loader)
-        if use_tqdm:
-            pbar.set_description(
-                (
-                    f'Epoch {e}, train loss: {train_loss:.5f} '
-                    f'train f error: {train_pino:.5f}; '
-                    f'data l2 error: {data_l2:.5f}'
-                )
-            )
-        if wandb and log:
-            wandb.log(
-                {
-                    'Train f error': train_pino,
-                    'Train L2 error': data_l2,
-                    'Train loss': train_loss,
-                }
-            )
-
-        if e % 100 == 0:
-            save_checkpoint(config['train']['save_dir'],
-                            config['train']['save_name'].replace('.pt', f'_{e}.pt'),
-                            model, optimizer)
-    save_checkpoint(config['train']['save_dir'],
-                    config['train']['save_name'],
-                    model, optimizer)
-    print('Done!')
 
 
 def train_2d_burger(model,
